Tried to add locks! Hard to test right now - ask for help?

+++ threaded-file-transfer-lab/threadServer.py
@@ -1,7 +1,7 @@
 #! /usr/bin/env python3
 
 # Server file transfer program. Based on emphaticDemo: see COLLABORATIONS for details.
-import sys, os, socket, params, time
+import sys, os, socket, params, time, threading
 from threading import Thread
 from framedSock import FramedStreamSock
 
@@ -30,10 +30,15 @@ class ServerThread(Thread):
     def __init__(self, sock, debug):
         Thread.__init__(self, daemon=True)
         self.fsock, self.debug = FramedStreamSock(sock, debug), debug
+        self.lock = threading.Lock()
         self.start()
         
     def run(self):
         
+        if debug: print(self.fsock, "Waiting for lock.")
+        
+        self.lock.acquire()  # We want whatever thread gets here first to claim the file.
+        if debug: print(self.fsock, "Lock acquired.")
         msg = self.fsock.receivemsg()  # First receive checks for error.
 
         if (msg == b"error"): # If you get an error message, stop!
@@ -46,7 +51,16 @@ class ServerThread(Thread):
             msg = self.fsock.receivemsg()
             
         filePath = os.getcwd() + "/server/" + fileName #Build path for file.
+
+        if os.path.isfile(filePath): # Don't allow overwrite of an already-existing file.
+            print(self.fsock, "file already exists! Stopping.")
+            self.lock.release() # Let go of the lock if you're done!
+            if debug: print (self.fsock, "lock released.")
+            return
+        
         myFile = open(filePath, 'wb')
+        self.lock.release() # First one here has claimed the file - let the other threads come through.
+        if debug: print(self.fsock, "relased lock.")
 
         while True:                    # Write to file until there is nothing to recieve!
             msg = self.fsock.receivemsg()
@@ -60,7 +74,6 @@ class ServerThread(Thread):
             time.sleep(0.001)
             ServerThread.requestCount = requestNum + 1
             msg = ("%s! (%d)" % (msg, requestNum)).encode()
-
 
 while True:
     sock, addr = lsock.accept()
